Renamed Runner to ScenarioRunner and moved to lib/
