+++ tests/test_yml_parsing.py
@@ -4,7 +4,7 @@ import unittest
 GMT_DIR = os.path.join(os.path.dirname(os.path.abspath(__file__)), '../')
 
 from lib import utils
-from runner import Runner
+from lib.scenario_runner import ScenarioRunner
 
 class TestYML(unittest.TestCase):
 
@@ -12,7 +12,7 @@ class TestYML(unittest.TestCase):
         test_root_file = 'tests/data/usage_scenarios/import_one_root.yml'
         name = 'test_' + utils.randomword(12)
 
-        runner = Runner(name=name, uri=GMT_DIR, uri_type='folder', filename=test_root_file)
+        runner = ScenarioRunner(name=name, uri=GMT_DIR, uri_type='folder', filename=test_root_file)
         runner.checkout_repository() # We need to do this to setup the file paths correctly
 
         runner.load_yml_file()
@@ -27,7 +27,7 @@ class TestYML(unittest.TestCase):
         test_root_file = 'tests/data/usage_scenarios/import_two_root.yml'
         name = 'test_' + utils.randomword(12)
 
-        runner = Runner(name=name, uri=GMT_DIR, uri_type='folder', filename=test_root_file)
+        runner = ScenarioRunner(name=name, uri=GMT_DIR, uri_type='folder', filename=test_root_file)
         runner.checkout_repository() # We need to do this to setup the file paths correctly
 
         runner.load_yml_file()
@@ -49,6 +49,6 @@ class TestYML(unittest.TestCase):
     def test_invalid_path(self):
         name = 'test_' + utils.randomword(12)
         test_root_file = 'tests/data/usage_scenarios/import_error.yml'
-        runner = Runner(name=name, uri=GMT_DIR, uri_type='folder', filename=test_root_file)
+        runner = ScenarioRunner(name=name, uri=GMT_DIR, uri_type='folder', filename=test_root_file)
         runner.checkout_repository() # We need to do this to setup the file paths correctly
         self.assertRaises(ValueError, runner.load_yml_file)
